Move DAP request authorization logic behind a callback
Currently, bearer tokens are included in the `DapTaskConfig` associated
with a task. This creates to limitations:

1. Bearer tokens need to be stored securely, whereas the task config
   does not have the same security requirements.
2. This codifies the sender authentication mechanism in
   draft-ietf-ppm-dap-01, which we expect to replace some time in the
   future.

To address these limitations, this change modifies `DapRequest` so that
the sender authentication mechanism is generic. It also introduces a new
role, `DapAuthorizedSender`, which implements the sender's functionality
for authentication. Likewise, the `DapAggregator` role has been extended
in rder to implement the receiver's functionality.