Parse Report extensions
Extensions are currently handled opaquely in a crate-private field
`ignored_extensions`. This change replaces this field with a public
`extensions` field whose type is a sequence of `Extension`s.

This is needed now in order to allow a `Report` to be constructed
outside of the crate. We will need this functionality eventually in
order to handle extensions properly.