Implement `dapf` CLI for interacting with deployments
Adds a new binary target, `dapf`, that implements a number of test
utilities. Initially this includes the following:
* upload: Generate a report and send it to the Leader
* collect: Send a collect request to the Leader
* collect-poll: Check if an aggregate result is ready and display it

This commit also adds a test script for demonstrating how `dapf` is
used.

In addition, while working on this change I noticed that we currently
don't authorize requests to collect URIs. I added a TODO to address
this.