daphne_worker: Improve collect URI handling
The current DAP draft (01) is ambiguous as to how long the Leader needs
to store aggregate results. But because the results are obtained via an
HTTP GET, the safest thing to do for now seems to be to store them
indefinitely. (The DAP spec allows the state to be explicitly dropped by
the Collector via an HTTP DELETE to the collect URI.)

Another issue is that the collect URI itself is easy for an HTTP client
to guess. And because this endpoint is not authenticated, a client can
force us to remove an aggregate result before the legitimate client
(i.e., the Collector) has a chance to retrieve it. This commit makes the
URI unpredictable by having it encode a pseeudorandom string.