daphne_worker: Process reports in order of arrival
Change report selector for `DaphneWorkerConfig` so that instead of
scheduling work per task and batch interval, reports are processed in
order of arrival.

A new DO is implemented, `LeaderAggregationJobQueue`, which maintaines
a queue of `ReportStore` instances. Once a report arrrives to an empty
`ReportStore` instance, an "aggregation job" is added to the queue; once
the instance becomes empty, the job is removed from the queue. Jobs are
ordered by the first report's time of arrival so that the oldest data is
processed first.

Other changes:

* daphne: Remove `task_id` from `DapLeader::get_pending_collect_jobs()`
  so that scheduling work on collect jobs doesn't have to depend on a
  particular task.

* Remove a TODO that was addressed by a previous change.