daphne_worker: Improve garbage collection
Adds a new durable object, `GarbageCollector`, that keeps track of every
DO instance that is created. When called, it invokes .delete_all() on
each DO instance scheduled for deletion. A DO instance schedules itself
for deletion by issuing a request to the garbage collector. This occurs
only once when the instance is first touched.

Replaces the `/internal/test/reset` endpoint with a new endpoint,
`/internal/delete_all`, that invokes the garbage collector. This allows
us to perform this task without needing to specify a range of buckets.
In particular, a batch interval and task ID are no longer needed.

This endpoint is only suitable for non-production deployments.
Eventually, we will schedule a cron job that periodically prunes the
oldest DO instances. In anticipation of this need, DO instances deleted
in the order in which they were touched.