`daphne`: Remove `worker` dependency
Remove `From<worker::Error` for `DapError`, thereby eliminating the
`worker` dependency from `daphne` altogether. Instead, implement
conversion explicitly in the `daphne_worker` crate.

This results in more verbose code, but the upside is that we no longer
need to synchronize the version of `worker` to use across the crates.
This is useful while workers-rs is under active development.