dapf, daphne: Add BatchSelector and Query structs
Define structs for controlling batch selection in DAP-02. Instead of a
batch interval, the Collector specifies a "query" in its collect
request; correspondingly, the Leader specifies a "batch selector" in its
aggregate share request.

This commit has no functional changes for the `daphne` crate. In
particular, only the "time-interval" query type is supported. However,
the input format for the `dapf` binary has changed.