Allow `HpkeDecrypter::get_hpke_config_for` to handle mutex guard
Change the return type of `HpkeDecrypter::get_hpke_config_for` to
`HpkeDecryptor::WrappedHpkeConfig: AsRef<HpkeConfig>`. This allows
Daphne-Worker to return a reference to an HPKE config protected by a
mutex.

For `DaphneWorkerConfig`, the return type is
`GuardedHpkeReceiverConfig<'a>`, which is required to live as long as
the mutex guard for the underlying data structure (a map from config IDs
to configs). The compiler does not know whether this object lives as
long as `DaphneWorkerConfig`, so it is necessary to add a lifetime to
the `HpkeDecrypter` trait.

Because `get_hpke_config_for()` is called at most once per HTTP request,
and because the HPKE config needs to be fetched from KV for each
request, this change does not save any overhead. (It doesn't add any
either.) However, since plan to move the DAP task configs to KV as part
of issue#82, this pattern will need to be replicated for
`DapAggregator::get_task_config_for()` as well.