dapne: Refactor roles test framework
In preparation for adding support for "fixed_size" queries, refactor the
test framework in roles_test.rs to make it easier to define per-test
task configurations. The main limitation of the current test code is
that the task list is hard-coded. By generating it at runtime we can
more easily define a set tasks that we want our code to cover.

This change also addresses another important limitation. All code used
to simulate interactions between the various parties is implemented on
`MockAggregator`. This makes testing quite awkward, particularly when
the parties need to share artifacts from dynamically generated tasks. To
address this, a struct `Test` is defined which stores all of the
artifacts associated with a logical "deployment".