daphne_worker: Generalize KV caching logic
HPKE configs are stored in KV, however they're cached in memory so that
we can avoid fetching from KV multiple times per request. In order to
implement draft-dcook-ppm-dap-interop-test-design-02 we will need to
store task configs in KV as well.

In preparation for this change, generalize the underlying caching so
logic so that it can be reused for tasks.