daphne_worker: Move bearer tokens to KV
Defines a new endpoint, /internal/test/add_authentication_token, for
storing bearer tokens per-task in KV. This obsoletes the variables
DAP_LEADER_BEARER_TOKEN_LIST and DAP_COLLECTOR_BEARER_TOKEN_LIST.

The map of task IDs to tokens is cached in memory. This table is guarded
by an RwLock. To avoid copying the task ID, we pass a reference to the
already parsed task ID to DapAggregator::authorized() by adding the task
ID field to the DapRequest structure. We still parse the task ID twice;
we should consider changing the DAP spec by moving the task IDs out of
message to make it easier to authorize the request before parsing it.