daphne_worker: Move task list to KV
Defines a new endpoint, /internal/test/add_task, for storing a task
config in KV. This obsoletes enviornment variable DAP_TASK_LIST.

In order to better align the JSON schema with the interop draft
(draft-dcook-ppm-dap-interop-test-design), move the min_batch_size
parameter of DapQueryConfig into DapTaskConfig.

We also take this opportunity to update the serialization of
VdafVerifyKey so that it is easier to deserialize. In particular, there
is no longer a need for the ShadowDapTaskConfig. The new format is more
verbose, but it is only relevant to our implementation.

Finally, change the DapAggregator::get_task_config() API so that we can
cache the task list in memory. To make this zoro-copy, we need to add an
explicit lifetime to indicate that the value returned lives as long as
the request.