Run `cargo upgrade --workspace` and fix any issues
Upgrading was impossible due to stale Janus dependencies. To resolve
this, remove these dependencies until we're ready to re-enable Janus
interop tests.

After successful upgrade, the compiler warns that we're using a
deprecated feature in the `wasm-bindgen` crate (deprecated as of 0.2.83).
`wasm-bindgen` is re-exported by `worker`, so it might be better to
upgrade there. For now, suprress the warnings by allowing use of the
deprecated function.