taskprov: Merge DpMechanism into DpConfig
Adding a new DP mechanism currently requires changing code in three
places: Adding a new DP_MECHANISM_* const, a new DpMechanims variant,
and whatever parameters are needed for DpConfig (currenntly no variants
of this structure are implemented). This change refactors serialization
of the DpConfig to make this a bit easeier.