taskprov: Clean up a couple minor things
Simplify verify key derivation slightly. Move up computation of `okm` in
the verify key derivation to preceed the match statement. The OKM
computation is identical for each of the branches, so this will avoid
having to de-duplicate when we add new variants. While at it, derive
`Copy` for `VdafVerifyKey` to get rid of the clone there.

While at it, correct a panic message that seems to expect that Poplar1
is the only other VDAF to implement.