daphne: Add plumbing for Prometheus metrics
Defines a struct `DaphneMetrics` where any metrics recorded while
processing or sending a DAP request will be enumerated. A registry is
required to construct this type. All of its members are crate-private.

Adds a new method to the `DapAggregator` that provides access to the
`DaphneMetrics` in use. This intended to be used by derived methods that
implement the protocol logic.