daphne_worker: Plumb state through request router
In preparation for pushing Prometheus metrics, plumb a reference to the
state of Daphne-Worker through the router so that we can scrape the
metrics before completing the HTTP request.

To accomplish this, we need a degree of separation between
`DaphneWorker`, used by the request router, and the state initialized
prior to creating the request router. The latter is encompassed with a
new struct, `DaphneWorkerState`, a reference to which is required to
construct `DaphneWorker`.