Implement `DapLeader::send_http_post()` for `MockAggregator`
This change adds a `peer: Option<Arc<MockAggregator` field to
`MockAggregator`. When the Leader is called on `send_http_post()`, it
resolves which `DapHelepr` to call on `peer` and issues the call
directly. This simulates an API call from the Leader to the Helper over
an open HTTP session.

To work around this missing method, currently the test harness rolls out
the code in `DapLeader::run_agg_job()` and
`DapLeader::run_collect_job()`. This change also refactors this code to
use these methods directly.