daphne_worker: Enforce key separation for HPKE across DAP versions
It's a good idea to ensure that secrets aren't reused across protocol
versions. Luckily, most secrets used by Daphne-Worker, are keyed by task
ID, which in turn is bound to a specific version. One exception are the
HPKE receiver keys.

To ensure key spearation for HPKE receiver keys, add the DAP version to
the KV names. That is, instead of

  hpke_receiver_config/{config_id}

the name is

  hpke_receiver_config/version/{version}/config_id/{config_id}

Relatedly:

* Check for unparsable version number in the list of KV keys for HPKE
  receiver configurations.

* change the string representation of `DapVersion::Unknown` to
  "unknown_version". This is to ensure that, if this string ends up in
  KV, it looks more obvious that something is wrong.

* Clean up the logic for parsing the KV name.

NOTE: This is a deployment-breaking change.