Spans should be traced at info level.
Previously we were tracing at "trace" level in
the mistaken belief this was required to avoid filtering, but
in fact the span level is just the level where the span will
get made, so we want the spans at info level as that means
they will exist and be useful for levels other than trace.

If we decided that the normal running state for the server
would have a higher level, e.g. ERROR, then we should increase
the span level too if we want to see it.