Only initialize state once per isolate instead of on every request.
As explained to us, a worker has a single thread of execution, but may
serve multiple requests, including concurrent requests, during the
lifetime of an isolate.

Example 1: Sequential requests.  A request arrives, the isolate serves
it, and the runtime leaves the isolate code loaded and ready.  If
another request arrives before the runtime decides to shutdown the
isolate, it can be served without reloading the worker.

Example 2: Concurrent (but not parallel) requests.  A request arrives,
the isolate starts serving it, and blocks on some I/O.  If another
request is available, the isolate may start working on that request.
There is only one thread of execution, and context switching occurs
only at defined yield points (like I/O).  So, there still needs to
be care taken due to concurrently executing threads, but they are
not pre-emptively scheduled or executing in parallel.

In both these cases, this optimization reduces the cost of handling
multiple requests during the life of one isolate substantially, as
creating the Daphne state from the worker environment's configuration
is costly.  It is safe to do this due to the nature of the data
and access patterns to the config.