daphne_worker: Clean up authorization logic
Currently either TLS client auth or a bearer token can be used to
authorize requests for taskprov tasks, but not both. This is problematic
from a deployment perspective, as we're likely to need to have both
methods in use while we transition to the new authorization method (TLS
client auth). This change modifies the authorization logic in
Daphne-Worker to validation all authorization methods presented and
authorize the request if all validate.

In addition, currently only one issuer/subject pair is trusted for the
certificate. In practice we're likely to issue certificates from
multiple issuers. This change replaces the single issuer/subject pair
with a sequence of trusted pairs.