Revert "Upgrade worker and serde-wasm-bindgen"
Downgrade worker 0.0.14 -> 0.0.12 and serde-wasm-bindgen -> 0.5.0 ->
0.4.5. In particular:

1. Revert commit 92903d8bdfdd121df588cffa1c9d44a95c2dde97

3. daphne_worker: Address an API-breaking change in worker

We observed bugs in workers-rs 0.0.14 that we could not reproduce locally
(wrangler dev), but only in Workers. We have not yet determined the root
cause, but it seemed to be associated with known bugs in 0.0.14.

The next version, 0.0.15, is out, but we will need to test it thoroughly
beefore we move to it.