draft02: Fix media type for AggregateContinueResp
The correct media type for this message is
"application/dap-aggregation-job-continue-resp"; we were incorrectly
returning "application/dap-aggregate-continue-resp". There two issues
here happening at once. First, we are overwriting the content type for
this message with the content type for draft04. Second, we are using the
incorrect content type for draft04. It should be
"application/dap-aggregation-job-resp".

This bug was introduced by the commit in which draft03 was replaced by
draft04. The code for resolving the media type for requests and
responses is quite complex. To make this sort of bug less likely in the
future, we replace the current string-matching functions with an
enumerated type that forces us to explicitly enumerate all the cases.

An enum is introduced, called `DapMediaType`, that replaces the
`Option<&'static str>` in the `DapRequest` and `DapResponse` structs.
Methods are defined on for decoding from and encoding to the
content-type HTTP header. These methods take the version, as is required
in order to properly resolve which representation of the media type is
needed.

All media types have a 1:1 mapping across the two versions, with one
exception: in draft02, there is a different media type for responses to
aggregate initialization and continuation requests, but in draft04,
there is no distinction made. For all other media types we prefer the
name in the latest draft.

Other minor things related to this change:

* This change requires adding the DAP version to the `DapResponse` so
  that the caller can figure out how to encode the media type.

* Strictly enforce the media type for all DAP requests. Doing so
  revealed a bug in the E2E test suite in which we failed to add the
  content-type header to certain requests.

The following unrelated changes are made:

* Replace `sender_for_media_type()` with a method on `DapMediaType` that
  not only resolves the sender for DAP requests, but the sender for DAP
  responses.

* tracing: Add a warning if the BearerToken denies a request due to
  unknown media type.

* editorial: Rename "Content-Type" to "content-type" in a few different
  places. The latter is the more canonical spelling.

* Improve error handling in various places.