dapf: Add command for generating an HPKE receiver config
Also, make `--task-id` optional, as it is not required for this command.

While at it, fix a runtime bug that resulted from a recent upgrade:

```
thread 'main' panicked at 'Cannot drop a runtime in a context where
blocking is not allowed. This happens when a runtime is dropped from
within an asynchronous context.',
/Users/christopherpatton/.cargo/registry/src/github.com-1ecc6299db9ec823/tokio-1.27.0/src/runtime/blocking/shutdown.rs:51:21
```

This is fixed by using the async versions of reqwest rather than the
blocking variants.