daphne_worker: Retry requests to HelperStateStore
Add a method, `with_retry()` on `DurableConnector` that, when invoked,
implements a simple linear-backoff retry mechanism. Further, retry
requests to HelperStateStore.

This change is intended to address transient errors in the Workers
platform that we noticed during load testing. In particular, we noticed
a non-negligible number of failures of requests to HelperStateStore, of
which there is a fresh instance per aggregation job.

Retries are safe for HelperStateStore because requests to this object
are idempotent. In general, whether retry is safe depends on the nature
of how the request changes the object's state.