Remove RwLockReadGuard usages
Locks in wasm are assumed to be single threaded without concurrency:
https://github.com/rust-lang/rust/blob/6f8c0557e0b73c73a8a7163a15f4a5a3feca7d5c/library/std/src/sys/unsupported/locks/rwlock.rs#L4

However, with Cloudflare Workers the situation is not so straightforward.

A single isolate can process multiple requests concurrently (not in
parallel), so holding ReadGuards across await points is not OK.

There are some official docs that refer to the behaviour, but I must
admit even having read those I wasn't convinced until I reproed it myself:
https://developers.cloudflare.com/workers/learning/how-workers-works/#distributed-execution

The errors only happen under load, likely because the Workers runtime
will readily spawn additional isolates.

The errors seen are:

    Error: The script will never generate a response.

    and

    Memory access out of bounds

I am reasonably confident this fixes the first, and somewhat confident
it fixes the latter.