daphne: Improve naming of protocol-level methods
The name `http_post_aggregate()` is misleading because it is not
necessarily used to handle an HTTP POST. In fact, in DAP-04, the initial
request is a PUT. Also, the name of the endpoints themselves have
changed in s way that is inconsistent across the two supported versions
(DAP-02 and -04). Similarly for other high-level API calls in
`roles.rs`.

This change updates the API and documentation to reflect the semantics
of the methods rather than the details of the HTTP request.