Do not clone request objects when making DO requests [#315].
Our prior technique made a request and then cloned it on each
attempt, never using the original.  This triggered a warning
from workerd on every DO request, as cloning prevents streaming
of the body.  (This wasn't important in our case as the body
is a small piece of JSON we just made, but workerd does not
know that.)  We now just make a new request object for each
attempt.