daphne_worker: Enusre HPKE config preference is respected
The current schema for the set of HPKE receiver configs stored in KV is
`HashMap<u8, HpkeReceiverConfig>`, e.g.:
```
{
    42 : {
        "config": {
            ...
        }
        "secret_key": ...
    },
    13: {
        "config" : {
            ...
        }
        "secret_key": ...
    }
}
```
When a Client asks for the HPKE config we currently return the "first"
element in this set. But since the set is unordered, there is no way
for the Aggregator to enforce a preference. For instance, it may prefer
P256 for the KEM.

In additon, once key rotation is implemented, the set will contain
stale configs that we are willing to use but don't want to advertise.

In order to ensure the value stored in KV reflects the servers
preference, replace the `HpkeReceiverConfigSet` type with
`HpkeReceiverConfigList`.