Coalesce report initialization operations into a trait
Define a trait `DapReportInitializer` that takes a sequence of reports
in the "consumed" state and transitions each report to the "initialized"
state. Implementers are expected to perform the early validation steps
(i.e., the protocol logic previously implemented by
`DapAggregator::check_early_reject()`) and initialize VDAF preparation.

The primary motivation for this change is to improve the performance of
`DaphneWorker`. VDAF preparation initialization is expensive, so we'd
like to parallelize the computation. Because the Workers runtime is
single-threaded, the best option is to offload VDAF preparation to
durable objects. In particular, `DaphneWorker`'s implementation of
`DapReportInitializer` piggy-backs this computation on the requests to
`ReportsProcessed`.

Co-authored-by: mendess <pmendes@cloudflare.com>