Wait to mark reports as aggregated until just before committing
Loads that require a lot of CPU time can trigger the Workers runtime to
reset itself, causing a 500 error. This is most likely to happen when
processing the AggregationJobInitReq, as this involves the expensive,
VDAF prep initialization step. Some reports may still get marked as
aggregated; if the peer retries the request, those reports will get
rejected.

This problem is so common that we need to make this request idempotent.
As a first step, wait to mark the reports as aggregated until just
before we have committed to aggregating them. In particular, instead of
doing this in `DapReportInitializer::initialize_reports()`, wait until
`DapAggregator::put_out_shares()`.