daphne_worker: Stabilize the PRF for internal use
We currently use cSHAKE128 for deriving ReportsPending and
ReportsProcessed shards and for computing the collection job ID for
DAP-02. This comes from `prio::vdaf::prg::PrgSha3`. This API is not
stable: the algorithm it uses depends on the VDAF version used, and we
have not settled on which XOF to use for VDAF. (In VDAF-07 we have
replaced cSHAKE128 with SHAKE128.) To avoid breaking changes for our
internal use cases, use `ring::hmac::HMAC_SHA256` instead.

The recommended key size for this PRF is 32 bytes. Accordingly, rotate
the keys in daphne_worker_test/wrangler.toml.