Provide more control when emitting a fatal error
When logging the error, instead of defaulting to the debug
representation of a type (using the `?` notation of tracing), give the
caller of the fatal_error macro control over which representation to
use. This mostly avoids strings from being debug printed unnecessarily
but also lets the caller pick between `?` (debug), `%` (display) or
nothing for when tracing knows how to handle the type.