vdaf: Make prep steps symmetric
In DAP-02, the Helper sends the Leader its prep share and the Leader
computes the prep message; in DAP-07, the Leader sends the Helper its
prep share and the Helper computes the prep message. To facilitate this,
update `vdaf::prio2` and `vdaf::prio3` so that either the Leader or the
Helper can compute the prep message.