draft07: Add ping-pong message framing as in draft-irtf-cfrg-vdaf-07
The payload of the PrepareInit message is prefixed by a `0` byte,
indicating to the consumer that this is the ping-pong "initialize"
message.

The payload of each transition in the AggregationJobResp is prefixed by
a `2` byte, indicating to the consumer that this is the ping-pong
"finish" message.

Since we only support 1-round VDAFs, it is not necessary to implement
the "continue" message.