Enable Prio3SumVecField64MultiproofHmacSha256Aes128
Define and enable a variant of Prio3SumVec variant that uses `Field64`
for the field, `XofHmacSha265Aes128` as the XOF, and allows an arbitrary
number of proofs.

The custom XOF uses a larger seed size than the standard one
(`XofTurboShake128`), so it is necessary to define a new variant of
`VdafVerifyKey` and other related structures.

This change also improves the unit tests in for the `vdaf::prio3` module
by replacing the Prio3-specific test code with the `AggregationJobTest`
runner.