taskprov: Enable Prio3SumVecField64MultiproofHmacSha256Aes128 in draft09
Enable the Prio3SumVec variant in taskprov. Don't enable it in draft02,
since the version of Prio3 we need (draft-irtf-cfrg-vdaf-08) is
incompatible with draft02.

While at it, unify the encoding logic across draft02 and draft09 so that
the same method is called to encode the query, VDAF, and DP parameters.
(The length prefix is only added in draft09.)