daphne_server: Plumb registered metrics through `App::new()`
In the downstream repo, the Oxy app will have two routers, both of which
try to register to the default prometheus registry. This causes a
runtime bug. To fix this, register the metrics once, then clone the
structure into each router.

While at it, align .gitignore with the downstream repo.