Add skeleton for Mastic integration
Add a new `VdafConfig` variant for the Mastic VDAF:
https://datatracker.ietf.org/doc/draft-mouris-cfrg-mastic/

We don't yet have a production-ready implementation of Mastic, so
implement a "dummy" (i.e., insecure) version that is only enabled for
tests and with feature "test-utils".

Mastic will be the first VDAF in Daphne that makes use of DAP's
"aggregation parameter". The aggregation parameter is used to refine the
input shares in a way required by the application. In case of Mastic,
the aggregation parameter defines the "candidate prefixes" in the case
of (weighted) heavy hitters or the "attributes" in case of
attribute-based metrics.

Because neither Prio2 nor Prio3 uses this feature, Daphne does not plumb
the aggregation parameter through the code the way we need to for
Mastic. This will require some refactoring. In the meantime, we stub
this code out in a way that let's us to test the `VdafConfig` variant
end-to-end.