Fix a bug in `VdafVerifyKey` decoding
`VdafConfig::get_decoded_verify_key()` has a bug that causes decoding to
fail for `Prio3SumVecField64MultiproofHmacSha256Aes128`. This was due to
an incorrectly handled enum variant.

To prevent this class of bug in the future, change the variants to
`L16` and `L32` for the 16-byte variant (Prio3 with the standard XOF)
and the 32-byte variant (Prio2 and Prio3 with the non-standard XOF).