Plumb aggregation parameter through the code
Replace the hard-coded aggregation parameter for testing Mastic with one
plumbed through the collection job. The change to the protocol code is
relatively minor, but it entails a major rework of the Leader
architecture.

The current architecture assumes there is no aggregation parameter,
hence aggregation can begin as soon as reports arrive. Aggregation job
creation is currently facilitated by `DapLeader::get_reports()`. The
Leader begins its processing loop (`leader::process()`) by calling this
function to get a set of reports to aggregate. It then issues a number
of aggregation jobs, then cycles through the queue of collection jobs to
complete any that can be completed.

This assumption no longer: with Mastic, the Leader must wait for a
collection job before it can begin scheduling aggregation jobs.

The new processing loop is centered around an object called the "work
queue". The work queue is a mixture of aggregation jobs and collection
jobs. The queue is manipulated with `DapLeader::enqueue_work()` and
`DapLeader::dequeue_work()`. The processing loop begins by dequeueing a
number of work items and processing them in order.

`MockAggregator` has been updated to wait to aggregate reports until a
collection job is created.

* Various minor improvements:

    * Clean up some buggy protocol logic around the creation of batches
      for fixed-size tasks. In particular, don't override the "current
      batch" query of the Collector's request.

    * Align input type of `DapAbort::batch_overlap()` with
      `query_mismatch()`

    * Add the batch ID to implementation of `Display` for `BatchSelector`

    * Implement `Display` for `Query`

    * Test multiple tasks in flight at once for `MockAggregator`

    * Align variable names for collection jobs with the latest draft

    * Resolve some TODOs