Align `daphne_server`'s Leader with new work queue semantics
Update `daphne_server` to match new the changes introduced by the
previous commit. Right now, `DapLeader` for `daphne_server::App` is used
for testing only. Thus, to keep this change as simple as possible,
refactor `daphne::MockAggregator` to share code with `daphne_server::App`.

As a result, most of the Leader's operating memory, including the work
queue, pending reports, and pending collection jobs, is no longer stored
in durable objects. This is good enough for now, but once it's time to
implement the production Leader, we'll need to devise a storage scheme,
using durable objects or something else.

Minor changes:

* Since removing the "report selector", the Leader can no longer control
  the "aggregation rate" in the way it did before.

* `DapLeader::put_report()` no longer tries to detect replays.
  Previously it was "best effort" and didn't catch all edge cases. (The
  definitive replay check happens when running an aggregation job.)

* Move `DapAggregator::current_batch()` to `DapLeader::current_batch()`
  (the Helper doesn't need to implement this)

* Add reports processed back to the telemtry returned by
  `leader::process()` (this was erroneously removed from the previous
  commit)

* Minor improvements to error handling in `MockAggregator`.