Prepare to support the next DAP version
Change `DapVersion::DraftLatest` to `DapVersion::Draft09` and use
`DapVersion::Latest` to track the next draft we'll target (currently
draft 10, which is not yet published). So far there are no
substantive changes in the next draft, but changes will be required to
enable Mastic in its "heavy hitters" mode of operation.