Remove Missing variant from DapMediaType.
Missing variants for an enum E should be encoded with `Option<E>`, this
way we can filter for missing once by filtering on None instead.

It also makes the code more obvious.

Another change is not using Option as a parameter when None is not
really a valid value of the type. This also makes the intentions of
the code clearer.