Simplify `DapReportInitializer`
Long ago, this trait was responsible for some of the early report
validation checks, including checking for replays and batch overlaps.
These days it is used solely for initializing preparation for a set of
reports.

However, `MockAggregator` currently does a bit more, in particular what
`DapAggregator::try_put_agg_share_span()` already does. After removing
this code, we can simplify the `DapReportInitializer` API so that it
does only what it's supposed to do.